requirements_bg.txt
rembg[gpu]
Pillow
